Brightcove refactor for web complete
