--- packages/brightcove-video/brightcove-video.stories.web.js
@@ -1,30 +1,0 @@
-import React from "react";
-import { storiesOf } from "@storybook/react-native";
-import BrightcoveVideo from "./src/brightcove-video.web";
-
-const policyKey =
-  "BCpkADawqM0NK0Rq8n6sEQyWykemrqeSmIQqqVt3XBrdpl8TYlvqN3hwKphBJRnkPgx6WAbozCW_VgTOBCNf1AQRh8KnmXSXfveQalRc5-pyNlSod5XzP99If2U";
-const accountId = "57838016001";
-const videoId = "4084164751001";
-const posterImageURI =
-  "https://clips.news.co.uk/thetimes/p5dzhoYzE6kYmndXxYdBsfnnyMzDVTyD/Ut_HKthATH4eww8X4xMDoxOmFkOxyVqc";
-
-const defaultVideoProps = {
-  policyKey,
-  videoId,
-  accountId,
-  poster: {
-    uri: posterImageURI
-  },
-  width: 320,
-  height: 180
-};
-
-storiesOf("Primitives/BrightcoveVideo", module).add(
-  "default player",
-  () => (
-    <BrightcoveVideo
-      {... defaultVideoProps}
-    />
-  )
-);
