--- packages/brightcove-video/src/brightcove-player.defaults.js
@@ -1,14 +1,0 @@
-export default {
-  width: 320,
-  height: 180,
-  playerId: "default",
-  paidonly: false,
-  onError: () => {},
-  onPlay: () => {},
-  onPause: () => {},
-  onFinish: () => {},
-  onProgress: () => {},
-  onDuration: () => {},
-  autoplay: false,
-  hideFullScreenButton: false
-};
