--- packages/brightcove-video/src/index.js
@@ -1,5 +1,0 @@
-import BrightcoveVideo from "./brightcove-video";
-import Splash_ from "./splash";
-
-export default BrightcoveVideo;
-export const Splash = Splash_;
